refactor(bcs): lift extract_bearer_token into bcs-auth-api
